refactor: remove src/voice/ from core — plugins own their own types (WOP-495)
Removes the voice module from core entirely. Voice plugins already define
their own types locally and should use the generic extension API
(registerExtension/getExtension) and capability registry instead of the
voice-specific registerSTTProvider/registerTTSProvider methods.

- Delete src/voice/ (index.ts, types.ts, registry.ts)
- Remove voice methods from WOPRPluginContext (types.ts, plugin-types/context.ts)
- Remove voice registry integration from context-factory.ts
- Replace legacy VoicePluginRequirements with manifest PluginRequirements
- Unify InstallMethod to single manifest definition
- Delete voice-registry.test.ts (tests deleted code)
- Remove dead STTProviderRef/TTSProviderRef interfaces
